Promote ov006/dScMgMCarlo_c to a single 23-function C++ TU
src_tu/actors/dScMgMCarlo_c.cpp -> src/actors/dScMgMCarlo_c.cpp, replacing
23 per-function legacy sources (18 flat-C, 5 already C++). The whole .text
range 0x020f7634..0x020f8e44 now reproduces from one object.

CONSOLIDATION IS WHERE THE WORK WAS. Each of the 23 legacy files compiled
in isolation with its own private shadow structs and extern declarations,
so they were free to contradict each other and never did so out loud. Put
in one TU they produced ~25 compile errors, every one a real disagreement
about the same object. Each was settled against evidence:

  - `Node` and the list-`Obj` are the SAME struct. Five partial views, all
    agreeing on every offset they share; unified into one layout with the
    fields commented.
  - A THIRD, unrelated `Obj` was not a list node at all -- it was a
    shape-only shadow of this scene class's OWN vtable, used for the
    slot-18 self-dispatch in InitResources. Revived as `SceneVtable` so
    the two meanings stop colliding. It becomes a real base-class virtual
    call once dScMgBase_c declares slots 18-35.
  - data_ov006_02142500 is the list HEAD and _02142504 the TAIL, proven by
    func_ov006_020f7740 walking head->next and tail->prev. Render starts
    from the tail, so both its loops step `prev`, not `next` -- corrobo-
    rated independently by func_ov006_020f8a3c's raw `*(char**)(r3 + 4)`.
  - _Z14ApproachLinearRiii and _Z15ApproachLinear2Rsss take REFERENCES;
    the mangled names say so, and a reference and a pointer pass the same
    address, so the accurate spelling costs no code.
  - func_ov006_020f8540 was declared (void) but called with an argument.
    ARM passes it in r0 and the body ignores it; spelling the parameter
    keeps the call site honest and changes no byte.
  - include/dScMgBase_c.h already declared data_ov004_020beb68 as void*,
    which settled that conflict against the TU's char*.

The destructor is defined inline in the class body, for the same reason as
dScMgBSC_c: out of line, mwcc emits the synthesized D0 ahead of the written
D1, and the cartridge has D1 first (0x020f7634) then D0 (0x020f76a8).
Inlining hands the key-function role to InitResources -- the next virtual
declared, non-inline, and defined here -- which is what emits
_ZTV13dScMgMCarlo_c. Slots 16 and 17 name D1 then D0, odr-using both, so
the compiler emits the pair in cartridge order, and the homeless D2 is gone.

decl_common.h loses the scalar declarations of the four list globals
(_021424fc/_02142500/_02142504/_02142508). They were arbitrary winners
among 19 contradictory isolated spellings; every consumer is in this TU,
which now declares all four with the recovered pointer type.

Verification:
  tubuild verify ov006/dScMgMCarlo_c   23/23 MATCH, objisolate clean,
                                       reloc-destinations clean,
                                       ROM-ascending section order
  tubuild linkcheck (pre-promotion)    SCRATCH-LINK-VERIFIED; 106/106
                                       exact; full ROM built and IDENTICAL
                                       to the stock build. dsd check
                                       symbols --fail still fails on the
                                       same 9 errors the baseline control
                                       has -- 0 new, so not attributable
                                       to this TU, but NOT green and not
                                       reported as such.
  rombuild -j16 (post-promotion)       11,088 source-built functions,
                                       reproducing 11,088, mismatching 0;
                                       106/106 exact, 100.000000%; the ROM
                                       it builds is sha256 d1506e90...,
                                       byte-identical to the stock build.
  layout_check / eligible / langmode / port_refcheck / dead-references: clean

All 13 compiler-emitted RTTI and vtable symbols are licensed in
compiler_only_output, every canonical address confirmed against
config/arm9 symbols.txt, and romdata_check byte-compares each against the
cartridge: 6 VERIFIED, 7 PARTIAL (exact prefixes), 0 DIFFER.
_ZTV13dScMgMCarlo_c reproduces 72 of the cartridge's 144 bytes -- 18 of 36
slots -- because dScMgBase_c still leaves slots 18-35 undeclared. That is a
verified prefix, not a disagreement, and widening it is the dScMgBase_c
follow-up, not this change.

The cartridge's own _ZTS strings corroborate the names: reading
extracted/dsd/arm9_overlays/ov006.bin at overlay base 34340544 gives
13dScMgMCarlo_c at 0x0213d594 and 19dScMgSingle3DBase_c at 0x0213bd00 --
so this TU emits no homeless typeinfo record.

Attribution: a 23-delete + 1-add collapse ends 22 git lineages, so
attribution.json takes 23 explicit override keys on the surviving path,
each carrying the first matcher the gate itself computes.
