+++ attribution.json
@@ -1040,6 +1040,29 @@
     "src/actors/dScMgBSC_c.cpp#func_ov006_02125364": "github-actions[bot]",
     "src/actors/dScMgBSC_c.cpp#_ZN10dScMgBSC_c6RenderEv": "github-actions[bot]",
     "src/actors/dScMgBSC_c.cpp#_ZN10dScMgBSC_c8BehaviorEv": "github-actions[bot]",
-    "src/actors/dScMgBSC_c.cpp#_ZN10dScMgBSC_c13InitResourcesEv": "github-actions[bot]"
+    "src/actors/dScMgBSC_c.cpp#_ZN10dScMgBSC_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#_ZN13dScMgMCarlo_cD1Ev": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#_ZN13dScMgMCarlo_cD0Ev": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7730": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7740": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7994": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7a00": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7a90": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7b10": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7b90": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7c10": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7e2c": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f7ee4": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8154": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8224": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f82d0": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8320": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f84a8": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8540": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#_ZN13dScMgMCarlo_c6RenderEv": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#_ZN13dScMgMCarlo_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8a3c": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#func_ov006_020f8c68": "github-actions[bot]",
+    "src/actors/dScMgMCarlo_c.cpp#_ZN13dScMgMCarlo_c13InitResourcesEv": "github-actions[bot]"
   }
 }
