+++ config/arm9/overlays/ov006/delinks.txt
@@ -4057,97 +4057,9 @@ src/MgMemoryMaster_Spawn.c:
     complete
     .text start:0x020f75d4 end:0x020f7634
 
-src/_ZN13dScMgMCarlo_cD1Ev.cpp:
+src/actors/dScMgMCarlo_c.cpp:
     complete
-    .text start:0x020f7634 end:0x020f76a8
-
-src/_ZN13dScMgMCarlo_cD0Ev.cpp:
-    complete
-    .text start:0x020f76a8 end:0x020f7730
-
-src/func_ov006_020f7730.c:
-    complete
-    .text start:0x020f7730 end:0x020f7740
-
-src/func_ov006_020f7740.cpp:
-    complete
-    .text start:0x020f7740 end:0x020f7994
-
-src/func_ov006_020f7994.c:
-    complete
-    .text start:0x020f7994 end:0x020f7a00
-
-src/func_ov006_020f7a00.c:
-    complete
-    .text start:0x020f7a00 end:0x020f7a90
-
-src/func_ov006_020f7a90.c:
-    complete
-    .text start:0x020f7a90 end:0x020f7b10
-
-src/func_ov006_020f7b10.c:
-    complete
-    .text start:0x020f7b10 end:0x020f7b90
-
-src/func_ov006_020f7b90.c:
-    complete
-    .text start:0x020f7b90 end:0x020f7c10
-
-src/func_ov006_020f7c10.c:
-    complete
-    .text start:0x020f7c10 end:0x020f7e2c
-
-src/func_ov006_020f7e2c.c:
-    complete
-    .text start:0x020f7e2c end:0x020f7ee4
-
-src/func_ov006_020f7ee4.cpp:
-    complete
-    .text start:0x020f7ee4 end:0x020f8154
-
-src/func_ov006_020f8154.c:
-    complete
-    .text start:0x020f8154 end:0x020f8224
-
-src/func_ov006_020f8224.c:
-    complete
-    .text start:0x020f8224 end:0x020f82d0
-
-src/func_ov006_020f82d0.c:
-    complete
-    .text start:0x020f82d0 end:0x020f8320
-
-src/func_ov006_020f8320.c:
-    complete
-    .text start:0x020f8320 end:0x020f84a8
-
-src/func_ov006_020f84a8.c:
-    complete
-    .text start:0x020f84a8 end:0x020f8540
-
-src/func_ov006_020f8540.c:
-    complete
-    .text start:0x020f8540 end:0x020f85b0
-
-src/_ZN13dScMgMCarlo_c6RenderEv.cpp:
-    complete
-    .text start:0x020f85b0 end:0x020f869c
-
-src/_ZN13dScMgMCarlo_c8BehaviorEv.cpp:
-    complete
-    .text start:0x020f869c end:0x020f8a3c
-
-src/func_ov006_020f8a3c.c:
-    complete
-    .text start:0x020f8a3c end:0x020f8c68
-
-src/func_ov006_020f8c68.c:
-    complete
-    .text start:0x020f8c68 end:0x020f8d08
-
-src/_ZN13dScMgMCarlo_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x020f8d08 end:0x020f8e44
+    .text start:0x020f7634 end:0x020f8e44
 
 src/func_ov006_020f8e44.cpp:
     complete
