+++ config/match_attempts.jsonl
@@ -930,7 +930,7 @@
 {"schemaVersion":1,"functionId":"arm9:0x0203e0ac","id":"arm9:0x0203e0ac","attemptId":"c2cd0c05ec2c413eaddff121787e9964","parentAttemptId":null,"module":"arm9","addr":33808556,"name":"func_0203e0ac","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_0203e0ac.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov002:0x020c2fec","id":"ov002:0x020c2fec","attemptId":"8387ab36ab0e402c91b7951aa1890774","parentAttemptId":null,"module":"ov002","addr":34353132,"name":"func_ov002_020c2fec","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov002_020c2fec.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020e39e0","id":"ov006:0x020e39e0","attemptId":"b000bed77d6145fbb99ff7aab1c16e09","parentAttemptId":null,"module":"ov006","addr":34486752,"name":"func_ov006_020e39e0","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020e39e0.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
-{"schemaVersion":1,"functionId":"ov006:0x020f8320","id":"ov006:0x020f8320","attemptId":"9f5906f11cbd43ca8b48d0b61e659d03","parentAttemptId":null,"module":"ov006","addr":34571040,"name":"func_ov006_020f8320","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020f8320.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
+{"schemaVersion":1,"functionId":"ov006:0x020f8320","id":"ov006:0x020f8320","attemptId":"9f5906f11cbd43ca8b48d0b61e659d03","parentAttemptId":null,"module":"ov006","addr":34571040,"name":"func_ov006_020f8320","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/actors/dScMgMCarlo_c.cpp","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020f9db8","id":"ov006:0x020f9db8","attemptId":"541d50383e1447ebb9d909484351bf03","parentAttemptId":null,"module":"ov006","addr":34577848,"name":"func_ov006_020f9db8","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020f9db8.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020faf6c","id":"ov006:0x020faf6c","attemptId":"62960718717c4395bae8522e59df8110","parentAttemptId":null,"module":"ov006","addr":34582380,"name":"func_ov006_020faf6c","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020faf6c.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov007:0x020c0eb0","id":"ov007:0x020c0eb0","attemptId":"9ac5fb0460b74716a3ba8850ae66243a","parentAttemptId":null,"module":"ov007","addr":34344624,"name":"func_ov007_020c0eb0","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov007_020c0eb0.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
@@ -940,7 +940,7 @@
 {"schemaVersion":1,"functionId":"arm9:0x0203e0ac","id":"arm9:0x0203e0ac","attemptId":"a51f29f9a41f43f492e7780208befd32","parentAttemptId":null,"module":"arm9","addr":33808556,"name":"func_0203e0ac","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_0203e0ac.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov002:0x020c2fec","id":"ov002:0x020c2fec","attemptId":"8805ac5e79074b4f9918565d0311ef00","parentAttemptId":null,"module":"ov002","addr":34353132,"name":"func_ov002_020c2fec","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov002_020c2fec.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020e39e0","id":"ov006:0x020e39e0","attemptId":"0f59c53c0fbd401ca1e19c5e88fdc4d1","parentAttemptId":null,"module":"ov006","addr":34486752,"name":"func_ov006_020e39e0","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020e39e0.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
-{"schemaVersion":1,"functionId":"ov006:0x020f8320","id":"ov006:0x020f8320","attemptId":"70253509ac8b49369ff983e2b090cadc","parentAttemptId":null,"module":"ov006","addr":34571040,"name":"func_ov006_020f8320","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020f8320.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
+{"schemaVersion":1,"functionId":"ov006:0x020f8320","id":"ov006:0x020f8320","attemptId":"70253509ac8b49369ff983e2b090cadc","parentAttemptId":null,"module":"ov006","addr":34571040,"name":"func_ov006_020f8320","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/actors/dScMgMCarlo_c.cpp","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020f9db8","id":"ov006:0x020f9db8","attemptId":"2f6fdf7a85b84d46ab8ad4950ac7d403","parentAttemptId":null,"module":"ov006","addr":34577848,"name":"func_ov006_020f9db8","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020f9db8.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov006:0x020faf6c","id":"ov006:0x020faf6c","attemptId":"389354be005b46c5a6db1185abc77191","parentAttemptId":null,"module":"ov006","addr":34582380,"name":"func_ov006_020faf6c","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov006_020faf6c.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
 {"schemaVersion":1,"functionId":"ov007:0x020c0eb0","id":"ov007:0x020c0eb0","attemptId":"cee9b1c29e5a4fdbbba5861fa3164507","parentAttemptId":null,"module":"ov007","addr":34344624,"name":"func_ov007_020c0eb0","status":"matched","kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code","author":"andrewboudreau","improvedNearMiss":false,"srcPath":"src/func_ov007_020c0eb0.c","label":"opus-fanout-2026-07-28","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":19}
