+++ config/match_provenance.jsonl
@@ -222,7 +222,7 @@
 {"id":"arm9:0x0203e0ac","module":"arm9","addr":33808556,"name":"func_0203e0ac","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_0203e0ac.c"}
 {"id":"ov002:0x020c2fec","module":"ov002","addr":34353132,"name":"func_ov002_020c2fec","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov002_020c2fec.c"}
 {"id":"ov006:0x020e39e0","module":"ov006","addr":34486752,"name":"func_ov006_020e39e0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov006_020e39e0.c"}
-{"id":"ov006:0x020f8320","module":"ov006","addr":34571040,"name":"func_ov006_020f8320","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov006_020f8320.c"}
+{"id":"ov006:0x020f8320","module":"ov006","addr":34571040,"name":"func_ov006_020f8320","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/actors/dScMgMCarlo_c.cpp"}
 {"id":"ov006:0x020f9db8","module":"ov006","addr":34577848,"name":"func_ov006_020f9db8","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov006_020f9db8.c"}
 {"id":"ov006:0x020faf6c","module":"ov006","addr":34582380,"name":"func_ov006_020faf6c","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov006_020faf6c.c"}
 {"id":"ov007:0x020c0eb0","module":"ov007","addr":34344624,"name":"func_ov007_020c0eb0","matchProvenance":{"kind":"ai","model":"claude-opus-5","reasoning":"high","harness":"claude-code"},"author":"andrewboudreau","srcPath":"src/func_ov007_020c0eb0.c"}
