--- src/_ZN13dScMgMCarlo_cD0Ev.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN13dScMgMCarlo_cD0Ev
-/* Real out-of-line definition, identical body to
-   _ZN13dScMgMCarlo_cD1Ev.cpp -- see that file's note. */
-#include "dScMgMCarlo_c.h"
-dScMgMCarlo_c::~dScMgMCarlo_c()
-{
-    __destroy_arr(mArray, 0x50, 0x30, (void *)func_ov006_020f7730);
-    func_ov006_020c1c64((char *)this + 0x4f38);
-}
