+++ src/actors/dScMgMCarlo_c.cpp
@@ -1,0 +1,1129 @@
+//cpp
+/* SHADOW translation unit -- generated by tools/tubuild.py create.
+ * ov006/dScMgMCarlo_c  (23 function(s))
+ *
+ * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
+ * build. It is a STARTING POINT (plan sec 7.3): local shadow declarations
+ * below were carried verbatim from the legacy files, not reconciled against
+ * real project headers -- that judgement call is left to a human/LLM review,
+ * the way pilot #1 reconciled PoleLift::Render and ::CleanupResources by hand
+ * (see notes/tu-reconstruction-pilot-report.md sec 5.2).
+ *
+ * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
+ * emits one .text section per function, in the REVERSE of source order, so
+ * the highest-address ROM function is written FIRST here. Do not reorder;
+ * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
+ * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
+ *
+ * Assembled from these legacy one-function sources (ROM address order):
+ *   [0] 0x020f7634  src/_ZN13dScMgMCarlo_cD1Ev.cpp
+ *   [1] 0x020f76a8  src/_ZN13dScMgMCarlo_cD0Ev.cpp
+ *   [2] 0x020f7730  src/func_ov006_020f7730.c
+ *   [3] 0x020f7740  src/func_ov006_020f7740.cpp
+ *   [4] 0x020f7994  src/func_ov006_020f7994.c
+ *   [5] 0x020f7a00  src/func_ov006_020f7a00.c
+ *   [6] 0x020f7a90  src/func_ov006_020f7a90.c
+ *   [7] 0x020f7b10  src/func_ov006_020f7b10.c
+ *   [8] 0x020f7b90  src/func_ov006_020f7b90.c
+ *   [9] 0x020f7c10  src/func_ov006_020f7c10.c
+ *   [10] 0x020f7e2c  src/func_ov006_020f7e2c.c
+ *   [11] 0x020f7ee4  src/func_ov006_020f7ee4.cpp
+ *   [12] 0x020f8154  src/func_ov006_020f8154.c
+ *   [13] 0x020f8224  src/func_ov006_020f8224.c
+ *   [14] 0x020f82d0  src/func_ov006_020f82d0.c
+ *   [15] 0x020f8320  src/func_ov006_020f8320.c
+ *   [16] 0x020f84a8  src/func_ov006_020f84a8.c
+ *   [17] 0x020f8540  src/func_ov006_020f8540.c
+ *   [18] 0x020f85b0  src/_ZN13dScMgMCarlo_c6RenderEv.cpp
+ *   [19] 0x020f869c  src/_ZN13dScMgMCarlo_c8BehaviorEv.cpp
+ *   [20] 0x020f8a3c  src/func_ov006_020f8a3c.c
+ *   [21] 0x020f8c68  src/func_ov006_020f8c68.c
+ *   [22] 0x020f8d08  src/_ZN13dScMgMCarlo_c13InitResourcesEv.cpp
+ */
+
+/* Includes: union of the legacy files', first-seen in ROM-ascending
+ * processing order. NOT verified for header ordering constraints (e.g. a
+ * common.h-before-X rule) -- watch for new compile errors after this. */
+#include "dScMgMCarlo_c.h"
+#include "types.h"
+#include "decl_common.h"
+
+/* Local shadow declarations carried from the legacy files verbatim.
+ * NOT reconciled against real project headers -- check include/*.h for
+ * each of these before compiling; a real header should usually win. */
+/* The minigame's board pieces live on one intrusive doubly-linked list.
+ * The 19 legacy files each modelled it partially and under two names
+ * ('Node' and 'Obj'); the offsets they agree on are unified here.
+ * data_ov006_02142500 is the head, _02142504 the tail (func_ov006_020f7740
+ * walks head->next and tail->prev), so Render, which starts from the tail,
+ * traverses via `prev`. */
+struct Node {
+    virtual void v0();          /* slot 0 -- per-piece render */
+    virtual void v1(int);       /* slot 1 -- per-piece update  */
+    /* 0x04 */ Node* prev;
+    /* 0x08 */ Node* next;
+    /* 0x0c */ int f_0c;        /* position x (12.4 fixed) */
+    /* 0x10 */ int f_10;        /* position y */
+    /* 0x14 */ int f_14;        /* target x   */
+    /* 0x18 */ int f_18;        /* target y   */
+    /* 0x1c */ int f_1c;        /* delta x    */
+    /* 0x20 */ int f_20;        /* delta y    */
+    /* 0x24 */ int f_24;
+    /* 0x28 */ short f_28;
+    /* 0x2a */ short f_2a;      /* board slot index */
+    /* 0x2c */ unsigned char f_2c;
+    /* 0x2d */ unsigned char f_2d;  /* state */
+    /* 0x2e */ unsigned char f_2e;
+};
+
+/* shadow typedef 'Fix12' */
+typedef int Fix12;
+
+/* Shape-only shadow of this scene class's OWN vtable, used for the slot-18
+ * self-dispatch in InitResources.  Not a distinct type -- delete once
+ * dScMgBase_c declares slots 18..35. */
+struct SceneVtable {
+    virtual void v0();  virtual void v1();  virtual void v2();  virtual void v3();
+    virtual void v4();  virtual void v5();  virtual void v6();  virtual void v7();
+    virtual void v8();  virtual void v9();  virtual void v10(); virtual void v11();
+    virtual void v12(); virtual void v13(); virtual void v14(); virtual void v15();
+    virtual void v16(); virtual void v17(); virtual void v18(int x);
+};
+
+/* shadow namespace 'GXS' */
+namespace GXS { void LoadOBJPltt(void const *, unsigned int, unsigned int); }
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Node', from the legacy file for func_ov006_020f7994, NOT applied:
+struct Node { char pad[8]; struct Node* next; };
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Node', from the legacy file for func_ov006_020f7b90, NOT applied:
+struct Node { char pad[8]; struct Node *next; char pad2[0x1e]; short key; };
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Obj', from the legacy file for func_ov006_020f8224, NOT applied:
+struct Obj
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Obj', from the legacy file for func_ov006_020f82d0, NOT applied:
+struct Obj { char pad[1]; };
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Obj', from the legacy file for func_ov006_020f8320, NOT applied:
+struct Obj {
+    char pad_00[0xc];
+    int f_0c;
+    int f_10;
+    int f_14;
+    int f_18;
+    int f_1c;
+    int f_20;
+    int f_24;
+    char pad_28[2];
+    short f_2a;
+    unsigned char f_2c;
+    unsigned char f_2d;
+};
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Node', from the legacy file for _ZN13dScMgMCarlo_c6RenderEv, NOT applied:
+struct Node {
+    virtual void f0();
+    Node* next;
+    char pad[0x18];
+    int field20;
+};
+*/
+
+/* TUBUILD CONFLICT -- alternate body of struct 'Obj', from the legacy file for _ZN13dScMgMCarlo_c13InitResourcesEv, NOT applied:
+struct Obj {
+    virtual void v0();  virtual void v1();  virtual void v2();  virtual void v3();
+    virtual void v4();  virtual void v5();  virtual void v6();  virtual void v7();
+    virtual void v8();  virtual void v9();  virtual void v10(); virtual void v11();
+    virtual void v12(); virtual void v13(); virtual void v14(); virtual void v15();
+    virtual void v16(); virtual void v17(); virtual void v18(int x);
+};
+*/
+
+extern "C" {
+extern int data_ov006_0213d5ac[];
+extern Node* data_ov006_021424fc;
+extern Node* data_ov006_02142500;
+extern Node* data_ov006_02142504;
+extern Node* data_ov006_02142508;
+extern short data_ov006_021424ec;
+extern int data_ov006_0213d570;
+extern int data_ov006_0213d574;
+extern void _Z14ApproachLinearRiii(int& x, int t, int s);
+extern void func_ov006_020f82d0(Node* o, int n);
+extern int RandomIntInternal(void* seed);
+extern int data_0209e650[];
+extern int data_ov006_0213d5e0[];
+extern int data_ov006_021424f4;
+extern int data_ov006_021424f8;
+extern int data_ov006_021424f0;
+extern int data_ov006_0213d568;
+extern int data_ov006_0213d56c;
+extern int data_ov006_0213d5e0[8];
+extern void func_ov006_020f84a8(char *o, int n);
+extern int func_ov006_020f7a00(void);
+extern int func_ov006_020f7b90(void);
+extern int data_020a0db0;
+extern unsigned short data_ov006_0213d600[];
+extern int data_ov006_0214250c[];
+extern u8 data_020a0e40[];
+extern u8 data_020a0de8[];
+extern u8 data_020a0de9[];
+extern u8 data_020a0dea[];
+extern u8 data_020a0deb[];
+extern int func_ov006_020f7b10(void);
+extern void Vec2_Sub(int* o, int* a, int* b);
+extern void func_0203d630(int *p, int m);
+extern int data_ov006_02133f18;
+extern void func_ov004_020b1ea4(int a, int b, int c, int d, int e, int f, int g);
+extern void Hud_RenderSprite(void* a0, int a1, int a2, int a3, int a4);
+void func_ov006_020c0aa8(char* p);
+void func_ov004_020b1bc8(char* a0, int a1, int a2, int a3);
+void func_ov004_020b6430(void);
+void func_ov004_020b1e34(char* a0, int a1, int a2, int a3);
+void func_ov006_020f8540(char* p);
+void func_ov006_020c1804(char* p);
+int func_ov006_020c1718(void* p);
+void func_ov004_020b65e4(void);
+void func_ov006_020c19d0(void* c);
+extern void func_ov006_0210a534(void *c);
+extern int GetGameLanguage(void);
+extern int LoadFile(int handle);
+extern void DecompressLZ16(void *src, void *dst);
+extern int func_ov006_020c1a88(void *p);
+extern unsigned char data_0209d45c;
+extern unsigned char data_0209d454;
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_0213d574, from the legacy file for func_ov006_020f7994, NOT applied: extern Fix12 data_ov006_0213d574; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142500, from the legacy file for func_ov006_020f7994, NOT applied: extern struct Node* data_ov006_02142500; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_021424fc, from the legacy file for func_ov006_020f7a90, NOT applied: extern int data_ov006_021424fc; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142508, from the legacy file for func_ov006_020f7a90, NOT applied: extern int data_ov006_02142508; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_021424fc, from the legacy file for func_ov006_020f7b10, NOT applied: extern int data_ov006_021424fc; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142508, from the legacy file for func_ov006_020f7b10, NOT applied: extern int data_ov006_02142508; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142500, from the legacy file for func_ov006_020f7b90, NOT applied: extern struct Node *data_ov006_02142500; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_021424fc, from the legacy file for func_ov006_020f7c10, NOT applied: extern int data_ov006_021424fc; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142508, from the legacy file for func_ov006_020f7c10, NOT applied: extern int data_ov006_02142508; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142500, from the legacy file for func_ov006_020f7c10, NOT applied: extern char *data_ov006_02142500; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142504, from the legacy file for func_ov006_020f8320, NOT applied: extern int data_ov006_02142504; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov006_02142504, from the legacy file for func_ov006_020f84a8, NOT applied: extern void* data_ov006_02142504; */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov006_020c0aa8, from the legacy file for _ZN13dScMgMCarlo_c13InitResourcesEv, NOT applied: extern void func_ov006_020c0aa8(void *p); */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov004_020beb68, from the legacy file for _ZN13dScMgMCarlo_c13InitResourcesEv, NOT applied: extern void *data_ov004_020beb68; */
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 22 -- _ZN13dScMgMCarlo_c13InitResourcesEv, 0x020f8d08, size 0x13c */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN13dScMgMCarlo_c13InitResourcesEv
+/* dScMgMCarlo_c::InitResources -- vtable slot 0.
+ *
+ * Attributed by tools/rtti_vtables.py --own dScMgMCarlo_c, this class's own
+ * slot 0 (fBase_c::InitResources). The old file's `recovered name:
+ * dScMgMCarlo_c_InitResources` agreed.
+ *
+ * The final `((Obj *)c)->v18(-1)` is a self-dispatch through this class's own
+ * vtable slot 18 -- left as a raw vtable-shim call, same shape the pre-migration
+ * file used, just through `this` instead of a `char *c` parameter. */
+s32 dScMgMCarlo_c::InitResources()
+{
+    char *c = (char *)this;
+    void *a;
+    void *b;
+
+    func_ov004_020b04d0(0x20);
+    func_ov006_0210a534(c);
+    a = (void *)LoadFile(data_ov006_0213d5b4[GetGameLanguage()]);
+    b = (void *)LoadFile(0xbb);
+    DecompressLZ16(a, (void *)0x6600000);
+    GXS::LoadOBJPltt(b, 0, 0x100);
+    Deallocate(a);
+    Deallocate(b);
+
+    data_0209d45c = 0x11;
+    data_0209d454 = 0x18;
+    *(unsigned short *)0x4000008 = (*(unsigned short *)0x4000008 & ~3) | 1;
+    *(unsigned short *)0x400100a = *(unsigned short *)0x400100a & ~3;
+    func_ov006_020c0aa8(c + 0x4660);
+    if (func_ov006_020c1a88(c + 0x4f38) == 0) return 0;
+
+    func_ov004_020b682c();
+    *(int *)(c + 0xa8) = func_ov004_020ad8b8();
+    *(int *)(c + 0xac) = *(int *)(c + 0xa8);
+    {
+        int v = func_ov004_020ad878();
+        if (data_ov004_020beb68) *(int *)((char *)data_ov004_020beb68 + 0xb4) = v;
+    }
+    ((SceneVtable *)c)->v18(-1);
+    this->unk_60aa = 0;
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 21 -- func_ov006_020f8c68, 0x020f8c68, size 0xa0 */
+/* -------------------------------------------------------------------------- */
+// @symbol func_ov006_020f8c68
+/* recovered: renamed to Class_Method, RTTI class fields named, declarations from a shared header */
+#include "decl_common.h"
+/* recovered: renamed to Class_Method, RTTI class fields named */
+#include "dScMgMCarlo_c.h"
+
+extern "C" {
+// recovered name: dScMgMCarlo_c_OnYoshiTryEat_020f8c68
+/* recovered: renamed to Class_Method */
+void func_ov006_020f7c10(char* p);
+void func_ov006_020c1604(char* c, int unused, short a2, int a3);
+void func_ov004_020b66d4(char* p);
+
+
+void func_ov006_020f8c68(char* c)
+{
+    struct dScMgMCarlo_c *self = (struct dScMgMCarlo_c *)(void *)c;
+    func_ov006_020f7c10(c + 0x51a8);
+    data_ov006_0213d564 = 0;
+    self->unk_60ae = 0;
+    self->unk_511e = 1;
+    func_ov006_020c1604(c + 0x4f38, 4, 4, (int)(c + 0x60ae));
+    self->unk_4f52 = 1;
+    self->unk_60aa = 0;
+    self->unk_60a8 = 1;
+    func_ov004_020b66d4(c + 0x6000);
+    data_ov004_020bc7d4 = 1;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 20 -- func_ov006_020f8a3c, 0x020f8a3c, size 0x22c */
+/* -------------------------------------------------------------------------- */
+#include "types.h"
+// @symbol func_ov006_020f8a3c
+/* recovered: renamed to Class_Method, RTTI class fields named, declarations from a shared header */
+#include "decl_common.h"
+/* recovered: renamed to Class_Method, RTTI class fields named */
+#include "dScMgMCarlo_c.h"
+
+extern "C" {
+// recovered name: dScMgMCarlo_c_OnTurnIntoEgg
+/* recovered: renamed to Class_Method */
+/* dScMgMCarlo_c::OnTurnIntoEgg - recovered from vtable slot identity */
+extern void FreeGfxSlotsById(int arg);
+extern void func_ov004_020b56c8(char* p);
+extern int _Z15ApproachLinear2Rsss(s16& r, s16 t, s16 s);
+
+extern unsigned short data_ov004_020bf9e4;
+extern void* data_ov004_020beb68;
+
+int func_ov006_020f8a3c(char* c)
+{
+    struct dScMgMCarlo_c *self = (struct dScMgMCarlo_c *)(void *)c;
+    switch (self->unk_60a8) {
+    case 4:
+        FreeGfxSlotsById(0x1d);
+        *(s16*)(((int)c + 0x60a8)) += 1;
+        break;
+    case 5:
+        if (func_ov006_020c1718(c + 0x4f38) != 0) {
+            int v = data_ov006_0213d570;
+            if (v != 0) {
+                func_ov004_020b5e40(v);
+                self->unk_60aa = -1;
+            } else {
+                func_ov004_020b56c8((char*)data_ov006_0213d568);
+            }
+            self->unk_60ae = 0;
+            *(s16*)(((int)c + 0x60a8)) += 1;
+        }
+        break;
+    case 6: {
+        void* q = data_ov006_02142500;
+        if (q != 0 && *(u8*)((char*)q + 0x2e) != 0) {
+            if (_Z15ApproachLinear2Rsss(*(s16*)(c + 0x60aa), 0, 1) != 0) {
+                char* r3;
+                int i;
+                int n;
+                self->unk_60aa = 0x10;
+                n = self->unk_60ae;
+                r3 = (char*)data_ov006_02142504;
+                for (i = 0; i < n; i++) {
+                    if (r3 == 0) break;
+                    r3 = *(char**)(r3 + 4);
+                }
+                if (r3 != 0) *(u8*)(r3 + 0x2e) = 0;
+                *(s16*)(((int)c + 0x60ae)) += 1;
+            }
+        } else {
+            if (data_ov004_020bf9e4 <= 1)
+                *(s16*)(((int)c + 0x60a8)) += 1;
+        }
+        break;
+    }
+    case 7: {
+        void* p = data_ov004_020beb68;
+        if ((p != 0 ? *(int*)((char*)p + 0xa8) : 0) != 0)
+            return 1;
+        func_ov006_020f7994();
+        self->unk_60aa = 0x1e;
+        *(s16*)(((int)c + 0x60a8)) += 1;
+        break;
+    }
+    case 8:
+        if (_Z15ApproachLinear2Rsss(*(s16*)(c + 0x60aa), 0, 1) != 0)
+            return 1;
+        break;
+    }
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 19 -- _ZN13dScMgMCarlo_c8BehaviorEv, 0x020f869c, size 0x3a0 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN13dScMgMCarlo_c8BehaviorEv
+/* dScMgMCarlo_c::Behavior -- vtable slot 6.
+ *
+ * Attributed by the ROM's vtable: the one slot where this class's table differs
+ * from dScMgSingle3DBase_c's. The old file's `recovered name:` comment agreed, and
+ * here it is right -- unlike the four the reference commit had to move up a level.
+ *
+ * The round's state machine, on the s16 at 0x60a8. Case 1 arms the shared
+ * dScMgBase_c counter pair at 0xc3/0xc4/0xc0 -- the identical idiom Flower's and
+ * Snowball's own Behaviors use, which is why include/dScMgBase_c.h names those
+ * three fields. It then falls through into case 2 deliberately.
+ *
+ * WAS A C99 FILE. Two consequences, both mechanical:
+ *   - the declarations move inside `extern "C"`, or they would mangle;
+ *   - `data_ov004_020beb68` is NOT redeclared here. include/dScMgBase_c.h already
+ *     declares it `void *`, and the old file's local `extern char *` would now be
+ *     a conflicting declaration rather than a private convenience. It is cast at
+ *     each use instead, which loads the same word.
+ * The local `typedef short s16;` is gone for the same reason: types.h arrives
+ * through decl_common.h and already has it. */
+s32 dScMgMCarlo_c::Behavior()
+{
+    void *arg = (void *)this;
+    unsigned char* c = (unsigned char*)arg;
+
+    switch (*(s16*)(c + 0x60a8)) {
+    case 1:
+        {
+            s16* p = (s16*)(((int)c + 0x60a8));
+            (*p)++;
+        }
+        if (c[0xc4] == 0) {
+            c[0xc3] = 1;
+            c[0xc4] = 1;
+            *(s16*)(c + 0xc0) = 0;
+        }
+        /* fall through */
+    case 2:
+        data_ov006_0213d574 = *(s16*)(c + 0x60ae) * 5 << 12;
+        if (*(s16*)(c + 0x60ae) == 4) {
+            if (func_ov006_020f7b10() == 0) {
+                *(s16*)(c + 0x60ae) = 0;
+                {
+                    s16* p = (s16*)(((int)c + 0x60a8));
+                    (*p)++;
+                }
+            }
+        }
+        break;
+    case 3:
+        if (func_ov006_020f7a90() == 0) {
+            int r5;
+            if (data_ov006_0213d564 == 1)
+                data_ov006_0213d564 = 0;
+            r5 = data_ov006_0213d574 >> 12;
+            if (data_ov006_0213d56c != 0 && r5 > 10 && r5 <= 18
+                && func_ov006_020c1718(c + 0x4f38) != 0) {
+                *(s16*)(c + 0x60ae) = 0;
+                *(s16*)(c + 0x511e) = 0;
+                func_ov006_020c1164(c + 0x4f38, 2, c + 0x60ae);
+                if (data_ov006_0213d56c == 2)
+                    *(s16*)(c + 0x4f52) = 0;
+            } else {
+                int lim = *(s16*)(c + 0x60ae) + 0x12;
+                if (r5 >= lim) {
+                    int flag = (data_ov006_021424fc != 0 && data_ov006_02142508 != 0);
+                    if (flag == 0)
+                        data_ov006_0213d574 = lim << 12;
+                }
+            }
+        } else {
+            if (data_ov006_0213d564 == 0) {
+                if (func_ov006_020f7b90() != 0) {
+                    data_ov006_0213d564 = 1;
+                } else {
+                    if (data_ov006_0213d570 != 0) {
+                        if (*(s16*)(c + 0x4f52) == 1) {
+                            *(s16*)(c + 0x4f52) = 0;
+                        } else if (func_ov006_020c16b4(c + 0x4f38) != 0) {
+                            {
+                                char *g = (char *)data_ov004_020beb68;
+                                if (g != 0) {
+                                    if (*(int *)(g + 0xb4) > 0)
+                                        *(int *)(((int)g + 0xb4)) -= 1;
+                                }
+                            }
+                            func_ov006_020c0d68(c + 0x4f38);
+                            func_ov004_020b0a54(5);
+                            {
+                                int d = data_ov006_0213d570;
+                                char *g = (char *)data_ov004_020beb68;
+                                int t = ((g != 0) ? *(int *)(g + 0xa8) : 0) - d;
+                                int u = (g != 0) ? *(int *)(g + 0xb4) : 0;
+                                func_ov004_020ad79c(t, u);
+                            }
+                            c[0xc3] = 0;
+                            {
+                                s16* p = (s16*)(((int)c + 0x60a8));
+                                (*p)++;
+                            }
+                        }
+                    } else {
+                        {
+                            char *g = (char *)data_ov004_020beb68;
+                            if (g != 0) {
+                                if (*(int *)(g + 0xb4) < 0x270f)
+                                    *(int *)(((int)g + 0xb4)) += 1;
+                                if (*(int *)(g + 0xb4) > *(int *)(g + 0xb8))
+                                    *(int *)(g + 0xb8) = *(int *)(g + 0xb4);
+                            }
+                        }
+                        func_ov004_020adb1c(*(int *)(c + 0xb4));
+                        func_ov006_020c0c80(c + 0x4f38);
+                        func_ov004_020b0a54(4);
+                        {
+                            char *g = (char *)data_ov004_020beb68;
+                            int t = (g != 0) ? *(int *)(g + 0xa8) : 0;
+                            func_ov004_020ad79c(data_ov006_0213d568 + t,
+                                                (g != 0) ? *(int *)(g + 0xb4) : 0);
+                        }
+                        c[0xc3] = 0;
+                        {
+                            s16* p = (s16*)(((int)c + 0x60a8));
+                            (*p)++;
+                        }
+                    }
+                    data_ov006_0213d564 = 0;
+                }
+            }
+        }
+        break;
+    }
+
+    func_ov004_020b65e4();
+    func_ov006_020c19d0(c + 0x4f38);
+    func_ov006_020f7740();
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 18 -- _ZN13dScMgMCarlo_c6RenderEv, 0x020f85b0, size 0xec */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN13dScMgMCarlo_c6RenderEv
+/* dScMgMCarlo_c::Render -- vtable slot 9.
+ *
+ * Attributed by tools/rtti_vtables.py --own dScMgMCarlo_c, this class's own
+ * slot 9 (dScMgBase_c::Render). The old file's `recovered name:
+ * dScMgMCarlo_c_Render` agreed. */
+s32 dScMgMCarlo_c::Render()
+{
+    char* c = (char*)this;
+    func_ov006_020c0aa8(c + 0x4660);
+    func_ov004_020b1bc8(c, 0xc, 0xc, 0);
+    func_ov004_020b6430();
+    func_ov004_020b1e34(c, 0xe0, 0x14, 1);
+
+    {
+        Node* n = data_ov006_02142504;
+        int i = 0;
+        for (;;) {
+            if (n == 0) break;
+            if (n->f_20 > 0) n->v0();
+            i++;
+            n = n->prev;
+            if (i >= 0x14) break;
+        }
+    }
+    {
+        int i = 0;
+        Node* n = data_ov006_02142504;
+        for (;;) {
+            if (n == 0) break;
+            if (n->f_20 == 0) n->v0();
+            i++;
+            n = n->prev;
+            if (i >= 0x14) break;
+        }
+    }
+    func_ov006_020f8540(c);
+    func_ov006_020c1804(c + 0x4f38);
+    return 1;
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 17 -- func_ov006_020f8540, 0x020f8540, size 0x70 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+/* The caller passes `this`; the ROM body never reads r0.  Spelling the
+ * parameter keeps the call site honest and costs no code. */
+void func_ov006_020f8540(char* p)
+{
+    int g = data_ov006_0213d56c;
+    if (g == 0) return;
+    func_ov004_020b1ea4(0xe8, 0x28, g, -1, -1, 0, 0);
+    Hud_RenderSprite((void*)data_ov006_02133f18, 0xe8, 0x18, -1, -1);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 16 -- func_ov006_020f84a8, 0x020f84a8, size 0x98 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f84a8(char* o, int n) {
+  *(short*)(o + 0x2a) = (short)n;
+  if (n >= 0x14) {
+    *(short*)(o + 0x28) = 1;
+  } else {
+    *(short*)(o + 0x28) = (short)((n % 5) * 2 + 1);
+  }
+  *(int*)(o + 0x24) = 0;
+  if (n == 0x13) data_ov006_02142504 = (Node*)o;
+  *(unsigned char*)(o + 0x2e) = 1;
+  *(int*)(o + 0x1c) = 0;
+  *(int*)(o + 0x20) = 0;
+  *(unsigned char*)(o + 0x2d) = 0;
+  *(int*)(o + 8) = 0;
+  *(int*)(o + 4) = *(int*)(o + 8);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 15 -- func_ov006_020f8320, 0x020f8320, size 0x188 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f8320(Node* o, int m)
+{
+    int tmp[3];
+
+    if (m >= 0x14) {
+        return;
+    }
+
+    if (o->f_2a >= 0x14) {
+        o->f_0c = (0x70 - ((data_ov006_0213d56c >> 2) << 1)) << 12;
+        o->f_10 = -0x30000;
+        data_ov006_02142504 = o;
+        o->f_2d = 1;
+    } else if (o->f_2d == 0) {
+        o->f_0c = (0x70 - ((data_ov006_0213d56c >> 2) << 1)) << 12;
+        o->f_10 = -0x30000;
+        o->f_2d = 4;
+        data_ov006_021424f4++;
+    } else {
+        o->f_2d = 4;
+        data_ov006_021424f4++;
+    }
+
+    o->f_14 = ((m % 5) * 32 + 0x30) << 12;
+    o->f_18 = ((m / 5) * 0x30) << 12;
+    o->f_2a = (short)m;
+
+    Vec2_Sub(tmp, &o->f_14, &o->f_0c);
+    o->f_1c = tmp[0];
+    o->f_20 = tmp[1];
+    func_0203d630(&o->f_1c, 0x124);
+
+    if (o->f_1c < 0) {
+        o->f_1c = -o->f_1c;
+    }
+    if (o->f_20 < 0) {
+        o->f_20 = -o->f_20;
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 14 -- func_ov006_020f82d0, 0x020f82d0, size 0x50 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f82d0(Node *o, int n) {
+    *(short*)((char*)o + 0x28) = (short)((4 - n % 5) * 2);
+    *(unsigned char*)((char*)o + 0x2d) = 5;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 13 -- func_ov006_020f8224, 0x020f8224, size 0xac */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f8224(Node *a, Node *b)
+{
+    int ai, bi, dm, dd;
+    if (b->f_2c != a->f_2c)
+        goto fail;
+    bi = b->f_2a;
+    ai = a->f_2a;
+    dm = ai % 5 - bi % 5;
+    dd = ai / 5 - bi / 5;
+    if (dm < 0)
+        dm = -dm;
+    if (dm >= 2)
+        goto fail;
+    if (dd < 0)
+        dd = -dd;
+    if (dd >= 2)
+        goto fail;
+    return 1;
+fail:
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 12 -- func_ov006_020f8154, 0x020f8154, size 0xd0 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f8154(Node *c)
+{
+    u8 idx;
+    int off;
+    int has;
+    int a, b;
+
+    if (data_ov006_0213d564 == 0) return 0;
+    if (func_ov006_020f7b10() != 0) goto fail;
+
+    idx = data_020a0e40[0];
+    off = idx * 4;
+    has = 0;
+    if (data_020a0de8[off]) {
+        if (data_020a0de9[off]) has = 1;
+    }
+    if (has == 0) goto fail;
+
+    a = data_020a0dea[idx * 4] - (c->f_0c >> 12);
+    b = data_020a0deb[idx * 4] - (c->f_10 >> 12);
+    if (a > 7 && a < 0x28 && b > 0 && b < 0x31) return 1;
+fail:
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 11 -- func_ov006_020f7ee4, 0x020f7ee4, size 0x270 */
+/* -------------------------------------------------------------------------- */
+/* func_ov006_020f7ee4 at 0x020f7ee4 (ov006), size 0x270
+ * flags: -O4,p -enum int -lang c++ -char signed -interworking -proc arm946e -gccext,on -msgstyle gcc
+ * Note: event must be passed through to func_ov006_020f8320 (zero-cost, r1 already
+ * holds it) - its liveness through case 0 is what pushes the case-0 temps up to
+ * r2/r3 as in the ROM. The u64-mask launder keeps the RMW address materialized.
+ */
+
+extern "C" {
+
+extern int data_ov006_0213d56c;
+extern short data_ov006_021424ec;
+extern int data_ov004_020bf9ec;
+extern int data_ov006_021424f4;
+
+extern unsigned int func_02012790(unsigned int a);
+extern void _ZN5Sound12PlayBank2_2DEj(unsigned int a);
+extern int func_0203d5dc(void *a, void *b);
+
+void func_ov006_020f7ee4(char *c, int event)
+{
+    switch (*(unsigned char *)(c + 0x2d)) {
+    case 0:
+        *(short *)((long long)(c + 0x28)) -= 1;
+        if (*(short *)(c + 0x28) != 0)
+            return;
+        data_ov006_0213d56c--;
+        func_ov006_020f8320((Node *)c, event);
+        return;
+
+    case 2:
+        if (*(short *)(c + 0x2a) != event) {
+            func_ov006_020f8320((Node *)c, event);
+            return;
+        }
+        if (func_ov006_020f8154((Node *)c) == 0)
+            return;
+        if (data_ov006_021424fc == 0) {
+            data_ov006_021424fc = (Node *)c;
+            *(unsigned char *)(c + 0x2d) = 3;
+            _ZN5Sound12PlayBank2_2DEj(0x153);
+            return;
+        }
+        if (func_ov006_020f8224((Node *)c, data_ov006_021424fc) != 0) {
+            data_ov006_02142508 = (Node *)c;
+            data_ov006_021424ec = 0x20;
+            *(unsigned char *)(c + 0x2d) = 3;
+            _ZN5Sound12PlayBank2_2DEj(0x154);
+            if (data_ov004_020bf9ec == 0)
+                data_ov004_020bf9ec = 1;
+            return;
+        }
+        func_02012790(0xe);
+        {
+            void *c2 = data_ov006_021424fc;
+            *((unsigned char *)c2 + 0x2d) = 2;
+        }
+        data_ov006_021424fc = 0;
+        return;
+
+    case 3:
+        if (func_ov006_020f8154((Node *)c) == 0)
+            return;
+        *(unsigned char *)(c + 0x2d) = 2;
+        if (data_ov006_021424fc != (void *)c)
+            return;
+        if (data_ov006_02142508 != 0)
+            return;
+        _ZN5Sound12PlayBank2_2DEj(0x155);
+        data_ov006_021424fc = 0;
+        return;
+
+    case 1:
+    case 4:
+        _Z14ApproachLinearRiii(*(int *)(c + 0xc), *(int *)(c + 0x14), *(int *)(c + 0x1c));
+        _Z14ApproachLinearRiii(*(int *)(c + 0x10), *(int *)(c + 0x18), *(int *)(c + 0x20));
+        _Z14ApproachLinearRiii(*(int *)(c + 0x24), 0x4000, 0x300);
+        if (func_0203d5dc((void *)(c + 0xc), (void *)(c + 0x14)) != 0)
+            return;
+        if (*(int *)(c + 0x24) != 0x4000)
+            return;
+        *(unsigned char *)(c + 0x2d) = 2;
+        _Z14ApproachLinearRiii(data_ov006_021424f4, 0, 1);
+        *(int *)(c + 0x1c) = 0;
+        *(int *)(c + 0x20) = 0;
+        return;
+
+    case 5:
+        if (_Z15ApproachLinear2Rsss(*(short *)(c + 0x28), 0, 1) == 0)
+            return;
+        _Z14ApproachLinearRiii(*(int *)(c + 0xc), -0x30000, 0x10000);
+        return;
+
+    default:
+        break;
+    }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 10 -- func_ov006_020f7e2c, 0x020f7e2c, size 0xb8 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void Hud_RenderSprite(void* a0, int a1, int a2, int a3, int a4);
+
+void func_ov006_020f7e2c(char *thiz)
+{
+    unsigned char k;
+    if (*(unsigned char*)(thiz + 0x2e) == 0) return;
+    k = *(unsigned char*)(thiz + 0x2d);
+    if (k == 0) return;
+    if (k == 3) {
+        if (((&data_020a0db0)[0] & 8) != 0) return;
+    }
+    {
+        int idx = (*(unsigned char*)(thiz + 0x2c) + 1) * 5
+                  + (*(int*)(thiz + 0x24) >> 12);
+        unsigned short e = data_ov006_0213d600[idx];
+        Hud_RenderSprite(
+            (void*)data_ov006_0214250c[e],
+            (*(int*)(thiz + 0xc) >> 12) + 0x18,
+            (*(int*)(thiz + 0x10) >> 12) + 0x18,
+            -1,
+            -1);
+    }
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 9 -- func_ov006_020f7c10, 0x020f7c10, size 0x21c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f7c10(char *c)
+{
+    int v;
+    char *o;
+    short i;
+    short j;
+    short off;
+    char *p;
+    char *q;
+    char *n2;
+    void *nx;
+
+    o = (char *)data_ov004_020beb68;
+    v = (o != 0) ? *(int *)(o + 0xb4) : 0;
+
+    if (v < 5) {
+        data_ov006_021424f8 = 6;
+        data_ov006_021424f0 = 6;
+        data_ov006_0213d568 = 5;
+    } else if (v < 0xa) {
+        data_ov006_021424f8 = 8;
+        data_ov006_021424f0 = 7;
+        data_ov006_0213d568 = 8;
+    } else {
+        data_ov006_021424f8 = 0xa;
+        data_ov006_021424f0 = 8;
+        data_ov006_0213d568 = 0xa;
+    }
+
+    data_ov006_0213d56c = data_ov006_021424f0 * data_ov006_021424f8;
+    data_ov006_0213d570 = data_ov006_021424f0 * data_ov006_021424f8;
+
+    do
+    {
+        for (i = 0; i < 8; i++)
+        {
+            if (i < data_ov006_021424f0)
+                data_ov006_0213d5e0[i] = data_ov006_021424f8;
+            else
+                data_ov006_0213d5e0[i] = 0;
+        }
+
+        j = 0;
+        p = c;
+        for (; j < 0x50; j++)
+        {
+            func_ov006_020f84a8(p, j);
+            if (j < data_ov006_0213d570)
+            {
+                p[0x2c] = (unsigned char)func_ov006_020f7a00();
+            }
+            p += 0x30;
+        }
+
+        off = 0;
+        data_ov006_021424fc = 0;
+        data_ov006_02142508 = 0;
+        data_ov006_021424ec = 0;
+        data_ov006_021424f4 = 0;
+        data_ov006_0213d574 = 0;
+        data_ov006_02142500 = (Node *)c;
+
+        if (data_ov006_0213d570 - 1 > 0)
+        {
+            q = c;
+            do
+            {
+                n2 = c + (off + 1) * 0x30;
+                nx = *(void **)(q + 8);
+                *(void **)(q + 8) = n2;
+                *(void **)(n2 + 4) = q;
+                while (*(void **)(n2 + 8) != 0)
+                {
+                    n2 = *(char **)(n2 + 8);
+                }
+                *(void **)(n2 + 8) = nx;
+                q += 0x30;
+                off++;
+            } while (off < data_ov006_0213d570 - 1);
+        }
+    } while (func_ov006_020f7b90() == 0);
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 8 -- func_ov006_020f7b90, 0x020f7b90, size 0x80 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f7b90(void)
+{
+    struct Node *p, *q;
+    p = data_ov006_02142500;
+    while (p != 0 && p->f_2a < 0x14) {
+        q = p->next;
+        while (q != 0 && q->f_2a < 0x14) {
+            if (func_ov006_020f8224(p, q) != 0) return 1;
+            q = q->next;
+        }
+        p = p->next;
+    }
+    return 0;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 7 -- func_ov006_020f7b10, 0x020f7b10, size 0x80 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f7b10(void)
+{
+    int ret = 1;
+    short t;
+    short f;
+    f = (short)(data_ov006_0213d574 >> 12);
+    t = (short)data_ov006_0213d570;
+    if (t > 20)
+        t = 20;
+    if (f == t)
+    {
+        if (data_ov006_021424f4 == 0)
+        {
+            if (data_ov006_021424fc == 0 || data_ov006_02142508 == 0)
+                ret = 0;
+        }
+    }
+    return ret;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 6 -- func_ov006_020f7a90, 0x020f7a90, size 0x80 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f7a90(void)
+{
+    int ret = 0;
+    short t;
+    short f;
+    f = (short)(data_ov006_0213d574 >> 12);
+    t = (short)data_ov006_0213d570;
+    if (t > 20)
+        t = 20;
+    if (f == t)
+    {
+        if (data_ov006_021424f4 == 0)
+        {
+            if (data_ov006_021424fc == 0 || data_ov006_02142508 == 0)
+                ret = 1;
+        }
+    }
+    return ret;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 5 -- func_ov006_020f7a00, 0x020f7a00, size 0x90 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+int func_ov006_020f7a00(void){
+    unsigned char pick = 0;
+    int total = 0;
+    int i;
+    int r;
+    for(i=0;i<8;i++) total += data_ov006_0213d5e0[i];
+    r = (int)(((unsigned int)RandomIntInternal(data_0209e650) & 0x7fffffff) >> 0x13);
+    total = (total * r) >> 0xc;
+    for(i=0;i<8;i++){
+        total -= data_ov006_0213d5e0[i];
+        if(total < 0){
+            pick = (unsigned char)i;
+            data_ov006_0213d5e0[i]--;
+            break;
+        }
+    }
+    return pick;
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 4 -- func_ov006_020f7994, 0x020f7994, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f7994(void){
+    struct Node* node = data_ov006_02142500;
+    short i = 0;
+    if((data_ov006_0213d574>>12) <= 0) return;
+    do {
+        if(node == 0) return;
+        func_ov006_020f82d0(node, i);
+        i = i+1;
+        node = node->next;
+    } while(i < (data_ov006_0213d574>>12));
+}
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 3 -- func_ov006_020f7740, 0x020f7740, size 0x254 */
+/* -------------------------------------------------------------------------- */
+extern "C" void func_ov006_020f7740(void) {
+    Node* head = data_ov006_02142500;
+    Node* n;
+    short i;
+
+    if (data_ov006_021424fc != 0) {
+        if (data_ov006_02142508 != 0) {
+            if (--data_ov006_021424ec == 0) {
+                if (data_ov006_021424fc->f_2a > data_ov006_02142508->f_2a)
+                    data_ov006_0213d574 = data_ov006_02142508->f_2a << 12;
+                else
+                    data_ov006_0213d574 = data_ov006_021424fc->f_2a << 12;
+
+                {
+                    Node* n1 = data_ov006_021424fc;
+                    if (data_ov006_02142500 == n1) data_ov006_02142500 = n1->next;
+                    if (data_ov006_02142504 == n1) data_ov006_02142504 = n1->prev;
+                    if (n1->prev != 0) n1->prev->next = n1->next;
+                    if (n1->next != 0) n1->next->prev = n1->prev;
+                    n1->next = 0;
+                    n1->prev = n1->next;
+                }
+                {
+                    Node* n2 = data_ov006_02142508;
+                    if (data_ov006_02142500 == n2) data_ov006_02142500 = n2->next;
+                    if (data_ov006_02142504 == n2) data_ov006_02142504 = n2->prev;
+                    if (n2->prev != 0) n2->prev->next = n2->next;
+                    if (n2->next != 0) n2->next->prev = n2->prev;
+                    n2->next = 0;
+                    n2->prev = n2->next;
+                }
+
+                data_ov006_02142508 = 0;
+                data_ov006_021424fc = 0;
+                _Z14ApproachLinearRiii(data_ov006_0213d570, 0, 2);
+            }
+            n = head;
+            for (i = 0; i < (data_ov006_0213d574 >> 12); i++) {
+                if (n == 0) return;
+                if (n->f_2d == 1) n->v1(i);
+                n = n->next;
+            }
+            return;
+        }
+    }
+
+    {
+        int t = (short)data_ov006_0213d570;
+        if (t > 0x14) t = 0x14;
+        _Z14ApproachLinearRiii(data_ov006_0213d574, t << 12, 0x800);
+    }
+    n = head;
+    for (i = 0; i < (data_ov006_0213d574 >> 12); i++) {
+        if (n == 0) return;
+        n->v1(i);
+        n = n->next;
+    }
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 2 -- func_ov006_020f7730, 0x020f7730, size 0x10 */
+/* -------------------------------------------------------------------------- */
+extern "C" {  /* .c-derived member: C linkage for the whole block */
+void func_ov006_020f7730(int *p) { p[0] = (int)data_ov006_0213d5ac; }
+}
+
+/* --------------------------------------------------------------------------
+ * ROM ordinals 0 and 1 -- _ZN13dScMgMCarlo_cD1Ev at 0x020f7634 (0x74)
+ *                     and _ZN13dScMgMCarlo_cD0Ev at 0x020f76a8 (0x88).
+ *
+ * Neither is written out here. ~dScMgMCarlo_c() is defined inline in the class
+ * body (include/dScMgMCarlo_c.h), so it cannot be this TU's key function;
+ * InitResources, the next virtual declared and non-inline, is, and this TU
+ * defines it. That emits _ZTV13dScMgMCarlo_c right here, whose slots 16 and 17
+ * name D1 and D0, odr-using both, so the compiler emits the pair out of line
+ * for us -- in cartridge order.
+ *
+ * Written out of line instead, mwcc emits the synthesized D0 AHEAD of the
+ * written D1, and linkcheck's pre-link audit refuses a TU whose licensed .text
+ * is not in ROM address order. Inlining also removes the homeless D2 that an
+ * out-of-line definition would emit with no ROM address to claim.
+ * -------------------------------------------------------------------------- */
