--- src/func_ov006_020f7730.c
@@ -1,2 +1,0 @@
-extern int data_ov006_0213d5ac[];
-void func_ov006_020f7730(int *p) { p[0] = (int)data_ov006_0213d5ac; }
